photos_org/requirements.txt
# 家庭单机版智能照片整理系统 - 依赖包列表
# 按照开发环境搭建指南的要求组织

# 核心Web框架
fastapi==0.104.1
uvicorn[standard]==0.24.0

# 数据库相关
sqlalchemy==2.0.23
aiosqlite==0.19.0

# 图像处理
Pillow==10.1.0
opencv-python==4.8.1.78
imagehash==4.3.2
numpy>=1.21.0,<2.0.0

# AI服务
dashscope==1.14.0

# 数据验证和序列化
pydantic==2.5.0
pydantic-settings==2.1.0

# 日志和配置
python-multipart==0.0.6
python-dotenv==1.0.0

# 开发和测试依赖
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# 代码质量工具
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# 可选依赖（用于增强功能）
# numpy==1.24.3  # 如果需要高级图像处理
# pandas==2.1.3  # 如果需要数据分析
# matplotlib==3.8.2  # 如果需要图像可视化

# 文档生成
mkdocs==1.5.3
mkdocs-material==9.4.8

# 性能监控（可选）
# psutil==5.9.6
# memory-profiler==0.61.0

# 版本信息
# Python >= 3.8
# 所有依赖包版本都经过测试，确保兼容性
